Add related articles to article pages
On blog article templates, in the bottom section with "Related articles" will show from now on.
It'll show article cards for articles which either have the same parent as the blog article that's being
read/visited or if the said article has no parents then it will show articles which have no parent as well.
Refs HEAL-12