Improve descriptions, help texts, error messages, etc
This is part of a big update to improve various texts in Shuup.

A lot of changes, touching almost every individual Shuup project's repo.

There are too many changes to try to describe them separately, but the main ones are:
1. Improve `help_text` descriptions. Fix typos, rewrite and change, write new ones.
2. Change some of the model field names (for example rename `available_from` -> `available_since`; `available_to` `available_until` - this is in a separate commit).
3. Improve some of the documentation.
4. Improve some in-code-comments.
5. Improve `settings.py` files. It's now clearer what each option does.
6. Change some of the naming.
7. Change the error/message system.

The main Shuup repo's Merge Request is here: https://github.com/shuup/shuup/pull/2113 (with a reasoning behind changes).